requirements.txt
fastapi 
uvicorn 
pydantic-settings 
pandas 
openpyxl 
requests 
python-dotenv
python-multipart
streamlit
openai